feat: shelf status persist on search
If a book is assigned to a shelf on the main page and that book appears
on the search page, the correct shelf is selected on the search page.
If that book's shelf is changed on the search page,
that change is reflected on the main page as well.
The option "None" should be selected if a book has not
been assigned to a shelf